+++ src/commands/mod.rs
@@ -1,0 +1,7 @@
+mod init;
+mod migrate;
+mod switch;
+
+pub use init::execute as init;
+pub use migrate::execute as migrate;
+pub use switch::execute as switch;